requirements.txt
pytest=6.2.4
pytest-html==3.1.1
pytest-xdist=2.3.0
Appium-Python-Client=1.2.0
